Move ordering to query, moved mapToCard to helpers. Refactored helpers
