+++ projects/helpers/contentful/client.ts
@@ -1,0 +1,9 @@
+import { createClient } from 'contentful';
+
+const space = process.env.NEXT_PUBLIC_CONTENTFUL_SPACE_ID;
+const accessToken = process.env.NEXT_PUBLIC_CONTENTFUL_ACCESS_TOKEN;
+
+export const client = createClient({
+  space: `${space}`,
+  accessToken: `${accessToken}`,
+});
