+++ projects/helpers/contentful/tempoString.ts
@@ -1,4 +1,5 @@
-import { createClient, Entry, EntryFields } from 'contentful';
+import { Entry, EntryFields } from 'contentful';
+import { client } from './client';
 
 export interface TempoEntry {
   title: EntryFields.Text;
@@ -6,14 +7,6 @@ export interface TempoEntry {
   tempoString: EntryFields.Text[];
   id: EntryFields.Text;
 }
-
-const space = process.env.NEXT_PUBLIC_CONTENTFUL_SPACE_ID;
-const accessToken = process.env.NEXT_PUBLIC_CONTENTFUL_ACCESS_TOKEN;
-
-const client = createClient({
-  space: `${space}`,
-  accessToken: `${accessToken}`,
-});
 
 export async function fetchDefaultEntry(): Promise<Entry<TempoEntry> | undefined> {
   return fetchEntry('default');
