+++ projects/portfolio/pages/blog/index.tsx
@@ -1,9 +1,9 @@
 import { Entry } from 'contentful';
 import { CardProps } from 'projects/component-library/stories/components/Card/Card';
 import { CardGrid } from 'projects/component-library/stories/components/CardGrid/CardGrid';
+import { BlogPost, fetchEntries } from 'projects/helpers/contentful/blogPost';
 import { LayoutContext } from 'projects/helpers/layoutContext';
-import { getGBDate } from 'projects/helpers/text';
-import { BlogPost, fetchEntries } from 'projects/portfolio/contentfulPosts';
+import { mapBlogPostToCard } from 'projects/helpers/mapToCard';
 import { useContext } from 'react';
 
 interface BlogProps {
@@ -33,14 +33,5 @@ const Blog = ({ data }: BlogProps): JSX.Element => {
 
   return <CardGrid cards={BlogPosts} size={300} />;
 };
-
-export function mapBlogPostToCard(blogPost: BlogPost): CardProps {
-  blogPost.endDate = getGBDate(blogPost.endDate);
-  return {
-    title: blogPost.title,
-    url: `blog/${blogPost.slug}`,
-    text: `${blogPost.endDate}\n${blogPost.excerpt}`,
-  };
-}
 
 export default Blog;
