+++ projects/portfolio/pages/tempo.tsx
@@ -1,7 +1,7 @@
 import { CheckboxList } from 'projects/component-library/stories/widgets/CheckboxList/CheckboxList';
 import { TextCard } from 'projects/component-library/stories/widgets/TextCard/TextCard';
+import { fetchDefaultEntry, TempoEntry } from 'projects/helpers/contentful/tempoString';
 import { LayoutContext } from 'projects/helpers/layoutContext';
-import { TempoEntry, fetchDefaultEntry } from 'projects/portfolio/contentfulTempo';
 import { useContext } from 'react';
 
 interface TempoProps {
